Add data persistence and initial fake data for Product model.
